Add implementation for transactions that does not require rollbacks as skeleton is always persisted using state unconfigured
