+++ src/integrationTest/java/uk/gov/hmcts/reform/wataskmanagementapi/services/CFTTaskDatabaseServiceTest.java
@@ -2,13 +2,16 @@ package uk.gov.hmcts.reform.wataskmanagementapi.services;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mockito.InjectMocks;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import uk.gov.hmcts.reform.wataskmanagementapi.SpringBootIntegrationBaseTest;
 import uk.gov.hmcts.reform.wataskmanagementapi.cft.entities.TaskResource;
 import uk.gov.hmcts.reform.wataskmanagementapi.cft.repository.TaskResourceRepository;
 
 import java.util.Optional;
 import java.util.UUID;
+import javax.persistence.EntityManager;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -20,7 +23,11 @@ class CFTTaskDatabaseServiceTest extends SpringBootIntegrationBaseTest {
     @Autowired
     TaskResourceRepository taskResourceRepository;
 
+    @InjectMocks
     CFTTaskDatabaseService cftTaskDatabaseService;
+
+    @MockBean
+    EntityManager entityManager;
 
     @BeforeEach
     void setUp() {
