+++ src/main/resources/application-local.yaml
@@ -14,4 +14,10 @@ logging:
     org.springframework:
       dao: info
       transaction: debug
+      jdbc.core.JdbcTemplate: info
+    org.hibernate:
+      SQL: info
+      type.descriptor.sql.BasicBinder: info
+      engine: info
+    uk.gov.hmcts.reform: debug
