+++ src/test/java/uk/gov/hmcts/reform/wataskmanagementapi/services/taskmanagementservicetests/GetTaskCountTest.java
@@ -1,0 +1,94 @@
+package uk.gov.hmcts.reform.wataskmanagementapi.services.taskmanagementservicetests;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.hmcts.reform.wataskmanagementapi.auth.permission.PermissionEvaluatorService;
+import uk.gov.hmcts.reform.wataskmanagementapi.config.LaunchDarklyFeatureFlagProvider;
+import uk.gov.hmcts.reform.wataskmanagementapi.controllers.request.SearchTaskRequest;
+import uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.camunda.CamundaSearchQuery;
+import uk.gov.hmcts.reform.wataskmanagementapi.services.CFTTaskDatabaseService;
+import uk.gov.hmcts.reform.wataskmanagementapi.services.CFTTaskMapper;
+import uk.gov.hmcts.reform.wataskmanagementapi.services.CamundaHelpers;
+import uk.gov.hmcts.reform.wataskmanagementapi.services.CamundaQueryBuilder;
+import uk.gov.hmcts.reform.wataskmanagementapi.services.CamundaService;
+import uk.gov.hmcts.reform.wataskmanagementapi.services.TaskManagementService;
+import uk.gov.hmcts.reform.wataskmanagementapi.taskconfiguration.services.ConfigureTaskService;
+import uk.gov.hmcts.reform.wataskmanagementapi.taskconfiguration.services.TaskAutoAssignmentService;
+
+import java.util.UUID;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+@ExtendWith(MockitoExtension.class)
+class GetTaskCountTest extends CamundaHelpers {
+
+    public static final String A_TASK_TYPE = "aTaskType";
+    public static final String A_TASK_NAME = "aTaskName";
+    @Mock
+    CamundaService camundaService;
+    @Mock
+    CamundaQueryBuilder camundaQueryBuilder;
+    @Mock
+    PermissionEvaluatorService permissionEvaluatorService;
+    @Mock
+    CFTTaskDatabaseService cftTaskDatabaseService;
+    @Mock
+    CFTTaskMapper cftTaskMapper;
+    @Mock
+    LaunchDarklyFeatureFlagProvider launchDarklyFeatureFlagProvider;
+    @Mock
+    ConfigureTaskService configureTaskService;
+    @Mock
+    TaskAutoAssignmentService taskAutoAssignmentService;
+    TaskManagementService taskManagementService;
+    String taskId;
+
+    @Test
+    void getTaskCount_should_succeed_and_return_count() {
+        SearchTaskRequest searchTaskRequest = mock(SearchTaskRequest.class);
+        CamundaSearchQuery camundaSearchQuery = mock(CamundaSearchQuery.class);
+        when(camundaQueryBuilder.createQuery(searchTaskRequest)).thenReturn(camundaSearchQuery);
+        when(camundaService.getTaskCount(camundaSearchQuery)).thenReturn(Long.valueOf(50));
+
+        long response = taskManagementService.getTaskCount(searchTaskRequest);
+
+        assertEquals(50, response);
+        verify(camundaService, times(1)).getTaskCount(camundaSearchQuery);
+    }
+
+    @Test
+    void getTaskCount_should_succeed_and_return_count_when_query_is_null() {
+        SearchTaskRequest searchTaskRequest = mock(SearchTaskRequest.class);
+        when(camundaQueryBuilder.createQuery(searchTaskRequest)).thenReturn(null);
+
+        long response = taskManagementService.getTaskCount(searchTaskRequest);
+
+        assertEquals(0, response);
+        verify(camundaService, times(0)).getTaskCount(any());
+    }
+
+    @BeforeEach
+    public void setUp() {
+        taskManagementService = new TaskManagementService(
+            camundaService,
+            camundaQueryBuilder,
+            permissionEvaluatorService,
+            cftTaskDatabaseService,
+            cftTaskMapper,
+            launchDarklyFeatureFlagProvider,
+            configureTaskService,
+            taskAutoAssignmentService
+        );
+
+        taskId = UUID.randomUUID().toString();
+    }
+}
+
